+++ include/ws_client/transport/HasSocketOperationsAsync.hpp
@@ -15,13 +15,26 @@ namespace ws_client
 {
 using byte = std::byte;
 
+template <typename S>
+concept CancelSlotLike = std::copy_constructible<S> && std::move_constructible<S>;
+
 /**
  * Concept for asynchronous socket template type parameters based on coroutines.
  * Requires the socket to support reading and writing bytes to the underlying socket, and closing the socket.
  * The functions MUST NOT throw exceptions, and instead return WSError object.
  */
-template <typename T, template <typename...> typename TTask>
-concept HasSocketOperationsAsync = requires(T t, std::span<byte> buffer, Timeout<>& timeout, bool fail_connection) {
+template <
+    typename T,                            // the socket type
+    template <typename...> typename TTask, // coroutine wrapper
+    typename TCancelSlot                   // cancellation type
+    >
+concept HasSocketOperationsAsync = CancelSlotLike<TCancelSlot> && requires(
+                                                                      T t,
+                                                                      std::span<byte> buffer,
+                                                                      Timeout<>& timeout,
+                                                                      bool fail_connection,
+                                                                      TCancelSlot& cancel
+                                                                  ) {
     /**
      * Checks if there is data available to be read from the socket without consuming it.
      * For SSL sockets, this checks for actual application data, not just SSL protocol bytes.
@@ -37,7 +50,7 @@ concept HasSocketOperationsAsync = requires(T t, std::span<byte> buffer, Timeout
      * 
      * @return The number of bytes read, or an error.
      */
-    { t.read_some(buffer, timeout) } -> std::same_as<TTask<std::expected<size_t, WSError>>>;
+    { t.read_some(buffer, timeout, cancel) } -> std::same_as<TTask<std::expected<size_t, WSError>>>;
 
     /**
      * Writes `buffer` to underlying socket.
@@ -45,7 +58,7 @@ concept HasSocketOperationsAsync = requires(T t, std::span<byte> buffer, Timeout
      * 
      * @return The number of bytes written, or an error.
      */
-    { t.write_some(buffer, timeout) } -> std::same_as<TTask<std::expected<size_t, WSError>>>;
+    { t.write_some(buffer, timeout, cancel) } -> std::same_as<TTask<std::expected<size_t, WSError>>>;
 
     /**
      * Shuts down socket communication.